requirements.txt
astroid==1.5.3
Django==1.11.7
djangorestframework==3.7.3
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
pytz==2017.3
six==1.11.0
wrapt==1.10.11
